closed network hierarchy identification system PRELIM
Code from Akshat Gokhale that can identify the list of URLs within a charter school’s website even if there is no directory hierarchy present (e.g., if the “About us” page is at school.com/Domain=459#x2i will the “Contact us” page is at school.com/Domain=g73kj&2). It does this by looking for all those URLs embedded in the given URL that share a common list of links—as in a menu—and therefore share with the given URL a “closed network”. This code needs to be reviewed for its usefulness and potentially adapted. THANKS AKSHAT FOR THIS HELP!